+++ app/src/main/java/net/wrappy/im/ui/ConversationListFragment.java
@@ -403,7 +403,7 @@ public class ConversationListFragment extends Fragment {
                     public void onClick(View v) {
                         Context context = v.getContext();
                         Intent intent = ConversationDetailActivity.getStartIntent(context);
-                        intent.putExtra("id", (int)chatId);
+                        intent.putExtra("id", chatId);
 //                        intent.putExtra("address", address);
                         intent.putExtra("nickname", nickname);
                         context.startActivity(intent);
